tests: move files around
