+++ tests/tests_integration/test_api/test_workflows.py
@@ -38,7 +38,7 @@ from cognite.client.data_classes.workflows import (
 from cognite.client.exceptions import CogniteAPIError
 from cognite.client.utils import timestamp_to_ms
 from cognite.client.utils._text import random_string
-from tests.tests_integration.test_api.test_simulators.seed_sim_data import (
+from tests.tests_integration.test_api.test_simulators.seed.resources import (
     ensure_workflow_simint_routine,
     finish_simulation_runs,
 )
